+++ example/lib/main.dart
@@ -163,7 +163,7 @@ void main() async {
               PlexDataCell.text("Grade"),
               PlexDataCell.text("Company"),
             ],
-            rows: List.empty(),//getTableData(),
+            rows: List.empty(), //getTableData(),
           ),
         ),
         PlexRoute(