+++ lib/plex_widget.dart
@@ -41,7 +41,7 @@ class PlexWidgetController<T> extends ChangeNotifier {
 
   void setValue(T? data) {
     this.data = data;
-    if(isDisposed) return;
+    if (isDisposed) return;
     notifyListeners();
   }